latest

+++ adminpage/views.py
@@ -3,9 +3,8 @@ from django.contrib.auth import authenticate, login, logout
 from forum.models import Forum
 
 def sign_in(request):
-	if request.user.is_authenticated:
-		print("authenticated")
-		return redirect('/')
+	if request.user.is_authenticated():
+		return redirect('admin_page')
 	context = {}
 	if request.method == 'POST':
 		username = request.POST.get('username')
@@ -13,39 +12,50 @@ def sign_in(request):
 		user = authenticate(username=username, password=password)
 		if user:
 			login(request, user)
-			return redirect('/')
-	else:
-		context['error'] = 'Invalid Username or Password'
-		context['username'] = username
+			return redirect('admin_page')
+	# else:
+	# 	context['error'] = 'Invalid Username or Password'
+	# 	context['username'] = username
 	return render(request, 'sign_in.html', context=context)
 
 def sign_out(request):
-    logout(request)
-    return redirect('sign_in')
+	    logout(request)
+	    return redirect('sign_in')
 
 def homepage(request):
-	forum = Forum.objects.filter(approved=False).order_by('-date')
-	return render(request, 'admin.html', {'forum': forum})
+	if request.user.is_authenticated():
+		forum = Forum.objects.filter(approved=False).order_by('-date')
+		return render(request, 'admin.html', {'forum': forum})
+	else:
+		error = "You are not logged in."
+		return render(request, 'sign_in.html', {'error': error})
 
 def approve(request):
-	if request.method == "POST":
-		_id = request.POST.get('requests_id', None)
+	if request.user.is_authenticated():
+		if request.method == "POST":
+			_id = request.POST.get('requests_id', None)
 
-		post_forum = Forum.objects.get(id=_id)
-		post_forum.approved = True
-		post_forum.save()
-		return redirect('admin_page')
-
+			post_forum = Forum.objects.get(id=_id)
+			post_forum.approved = True
+			post_forum.save()
+			return redirect('admin_page')
+		else:
+			return HttpResponse('Request must be POST.')
 	else:
-		return HttpResponse('Request must be POST.')
+		error = "You are not logged in."
+		return render(request, 'sign_in.html', {'error': error})
 
 def delete(request):
-	if request.method == "POST":
-		_id = request.POST.get('requests_id', None)
+	if request.user.is_authenticated():
+		if request.method == "POST":
+			_id = request.POST.get('requests_id', None)
 
-		post_forum = Forum.objects.get(id=_id)
-		post_forum.delete()
-		return redirect('admin_page')
+			post_forum = Forum.objects.get(id=_id)
+			post_forum.delete()
+			return redirect('admin_page')
 
+		else:
+			return HttpResponse('Request must be POST.')
 	else:
-		return HttpResponse('Request must be POST.')
+		error = "You are not logged in."
+		return render(request, 'sign_in.html', {'error': error})
